Backend/Frontend Cart Functionality
I added a backend cartRoute.js, for the front end to access the backend, as well as a Backend cartModel.js for adding, deleting and updating cart. Adding works, not sure on the other two. On the Frontend I added an add to cart button in the details.js file so that I could test functionality. it successfully adds to the cart table and increments items in your cart.