Fixed the cart count
Fixed the cart count, but discovered an error with the sessionID management. The Session ID changes constantly throughout the use of the app, when clicking on links etc. This is causing a problem with our cart when first launching the app as the cart ties itself to the session ID.