Updated Cart Functionality
I added a total, and a check out button. when you push the checkout button it confirms your order and then redirects to the Checkout page. this would need to be a form page that I have not gotten to yet! We will need post route in the backend to store the order information. I'm not entirely sure if this will work or not! Just getting some headway for checckout!