ReAdded the Search Functionality
I readded Braedons stuff, but then I tried to mess with the order.js and order.css file so that it looks better, but I cna't get it. Found an issue with the search function. the Book Details does not load when we search by author or genre, but does load by title